+++ src/createinstance.hpp
@@ -27,6 +27,10 @@ class CreateInstance : public QDialog
         bool testPort(const quint16& port);
         quint16 genPort();
 
+        void restartServer(const QString& name, const QString& gameName,
+                           const quint16& port, const bool& useUPNP,
+                           const bool& isPublic);
+
     private slots:
         void on_servers_currentTextChanged(const QString& arg1);
         void on_portNumber_textChanged(const QString& arg1);